intro_workshop/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.51.0
aws-cdk.aws-apigateway==1.51.0
aws-cdk.aws-applicationautoscaling==1.51.0
aws-cdk.aws-autoscaling-common==1.51.0
aws-cdk.aws-certificatemanager==1.51.0
aws-cdk.aws-cloudformation==1.51.0
aws-cdk.aws-cloudwatch==1.51.0
aws-cdk.aws-dynamodb==1.51.0
aws-cdk.aws-ec2==1.51.0
aws-cdk.aws-efs==1.51.0
aws-cdk.aws-elasticloadbalancingv2==1.51.0
aws-cdk.aws-events==1.51.0
aws-cdk.aws-iam==1.51.0
aws-cdk.aws-kms==1.51.0
aws-cdk.aws-lambda==1.51.0
aws-cdk.aws-logs==1.51.0
aws-cdk.aws-route53==1.51.0
aws-cdk.aws-s3==1.51.0
aws-cdk.aws-s3-assets==1.51.0
aws-cdk.aws-sns==1.51.0
aws-cdk.aws-sqs==1.51.0
aws-cdk.aws-ssm==1.51.0
aws-cdk.cloud-assembly-schema==1.51.0
aws-cdk.core==1.51.0
aws-cdk.custom-resources==1.51.0
aws-cdk.cx-api==1.51.0
aws-cdk.region-info==1.51.0
boto3==1.14.20
botocore==1.17.20
cattrs==1.0.0
cdk-dynamo-table-viewer==3.1.2
constructs==3.0.4
docutils==0.15.2
jmespath==0.10.0
jsii==0.21.2
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.9
